added error.hpp
The error.hpp header contains exception handling code. It defines a base gladus::error class which derives from std::exception. Based on this an opengl_error is defined that captures an error returned from glGetError(); and runtime_error which captures a nested chain of runtime_errors and one underlying opengl_error.

Additional functions are provided that help to check for OpenGL errors, or throw an exception if an OpenGL error is encountered. Two macros, if_opengl_error and incase_opengl_error, are provided to shorten OpenGL error handling code. The error checking functions may be converted into a nop via a preprocessor define.